+++ src/sort/bubbleSort/index.ts
@@ -2,7 +2,7 @@ import { swap, Comparator, compareFunctionType } from '../../utils';
 
 export function bubbleSort<T = unknown>(originalArray: T[], compareCallback?: Comparator | compareFunctionType): T[] {
   const comparator = new Comparator(compareCallback);
-  for (let i = 0; i < originalArray.length; ++i) {
+  for (let i = 0; i < originalArray.length - 1; ++i) {
     let isSwap = false;
     for (let j = 0; j < originalArray.length - i - 1; ++j) {
       if (comparator.greaterThan(originalArray[j], originalArray[j + 1])) {
@@ -11,7 +11,7 @@ export function bubbleSort<T = unknown>(originalArray: T[], compareCallback?: Co
       }
     }
     if (!isSwap) {
-      break;
+      return originalArray;
     }
   }
   return originalArray;