+++ src/sort/quickSort/quickSortExchange.ts
@@ -1,26 +1,26 @@
 import { Comparator, compareFunctionType } from '../../utils';
 
 export function quickSortExchange<T = unknown>(originalArray: T[], comparator: Comparator | compareFunctionType): T[] {
-  comparator = new Comparator(comparator);
   if (originalArray.length <= 1) {
     return originalArray;
   }
 
-  const pivotElement = originalArray[originalArray.length >>> 1];
-  const center = [];
+  const newComparator = new Comparator(comparator);
+  const baseItem = originalArray[originalArray.length >>> 1];
   const left = [];
+  const center = [];
   const right = [];
 
   for (let i = 0; i < originalArray.length; i++) {
-    if (comparator.equal(originalArray[i], pivotElement)) {
+    if (newComparator.equal(originalArray[i], baseItem)) {
       center.push(originalArray[i]);
-    } else if (comparator.lessThan(originalArray[i], pivotElement)) {
+    } else if (newComparator.lessThan(originalArray[i], baseItem)) {
       left.push(originalArray[i]);
     } else {
       right.push(originalArray[i]);
     }
   }
 
-  return quickSortExchange(left, comparator)
-    .concat(center, quickSortExchange(right, comparator));
+  return quickSortExchange(left, newComparator)
+    .concat(center, quickSortExchange(right, newComparator));
 }