+++ src/sort/quickSort/quickSortInPlace.ts
@@ -1,16 +1,16 @@
 import { Comparator, swap, compareFunctionType } from '../../utils';
 
 export function quickSortInPlace<T = unknown>(originalArray: T[], compareCallback?: Comparator | compareFunctionType, left = 0, right: number = originalArray.length - 1): T[] {
-  const comparator = new Comparator(compareCallback);
   if (originalArray.length > 1 && left < right) {
-    const partitionIndex = partitionArray(originalArray, comparator, left, right);
+    const comparator = new Comparator(compareCallback);
+    const baseIndex = partitionArray(originalArray, comparator, left, right);
 
-    if (left < partitionIndex - 1) {
-      quickSortInPlace(originalArray, comparator, left, partitionIndex - 1);
+    if (left < baseIndex - 1) {
+      quickSortInPlace(originalArray, comparator, left, baseIndex - 1);
     }
 
-    if (partitionIndex < right) {
-      quickSortInPlace(originalArray, comparator, partitionIndex, right);
+    if (baseIndex < right) {
+      quickSortInPlace(originalArray, comparator, baseIndex, right);
     }
   }
 
@@ -18,16 +18,16 @@ export function quickSortInPlace<T = unknown>(originalArray: T[], compareCallbac
 }
 
 function partitionArray(originalArray, comparator: Comparator, left: number, right: number): number {
-  const pivotElement = originalArray[(left + right) >>> 1];
+  const baseItem = originalArray[(left + right) >>> 1];
   let i = left;
   let j = right;
 
   while (i <= j) {
-    while (comparator.lessThan(originalArray[i], pivotElement)) {
+    while (comparator.lessThan(originalArray[i], baseItem)) {
       i++;
     }
 
-    while (comparator.greaterThan(originalArray[j], pivotElement)) {
+    while (comparator.greaterThan(originalArray[j], baseItem)) {
       j--;
     }